County Level Vaccination Data

Vaccination data from the California Department of Public Health. The data found online include the vaccination information reported by schools for each school year. This dataset is from the 2017-2018 school year. Dataset contains information for each county in CA. To compile county level data, numbers for all schools in each county are averaged.

Format

A data frame with variables:

Jurisdiction

Jurisdiction of California.

total_Enrollment

Total number of students enrolled.

total_Up_to_Date

Total number of up to date students. Up to date students must have four doses of polio vaccine (three doses are acceptable if at least one dose was received on or after the fourth birthday), five DTaP/DTP/DT vaccine doses (four doses are acceptable if at least one dose was received on or after the fourth birthday), and two doses of measles-containing vaccine, at least one of which must be measles, mumps, and rubella combined. Both doses must have been received on or after the first birthday. In addition, student must have three doses of hepatitis B vaccine and one dose of varicella vaccine or physician-documented varicella disease.

total_Conditional

Total number of students that do not meet all requirements. This may because the student lacks but is not yet due for a required dose, has a physician affidavit of Temporary Medical Exemption for one or more doses, or is a transfer student who has no record yet.

total_PME

Total number of students with a Permanent Medical Exemption (PME). Indicates that the physical condition of the student or medical circumstances relating to the student are such that immunization is permanently not indicated.

total_PBE

Total number of students with a Personal Belief Exemption (PBE), whereby a parent requests exemption from the immunization requirements for school entry because all or some immunizations are contrary to the parent's beliefs.

total_DTP

Total number of students with full DTP vaccination course. Full course defined as 4 or more doses of any diphtheria and tetanus toxoids and pertussis vaccines including diphtheria and tetanus toxoids, and any acellular pertussis vaccine (DTaP/DTP/DT).

total_Polio

Percent of students with full Polio vaccination course. Full course defined as 3 or more doses of polio vaccine; oral polio vaccine (OPV) or inactivated polio vaccine (IPV) or any combination of these.

total_MMR

Total number of students with full MMR vaccination course. Full course defined as 2 doses of measles-mumps-rubella vaccine.

total_HepB

Total number of students with full HepB vaccination course. Full course defined as 3 or more doses of Hepatitis B vaccine.

total_Var

Total number of students with full Varicella vaccination course. Full course defined as 1 or more doses of varicella at or after child's first birthday, unadjusted for history of varicella illness.

percent_Up_to_Date

The percent of up to date students. See total_Up_to_Date.

percent_Conditional

The percent of students that are conditional. See total_Conditional.

percent_PME

The percent of students with PME. See total_PME.

percent_PBE

The percent of students with PBE. See total_PBE.

percent_DTP

The percent of students vaccinated with DTP. See total_DTP.

percent_Polio

The percent of students vaccinated against polio. See total_Polio.

percent_MMR

The percent of students vaccinated with MMR. See total_MMR.

percent_HepB

The percent of students vaccinated against HepB. See total_HepB.

percent_Var

The percent of students vaccinated against varicella. See total_Var.

Year

The year that the data was collected. School years represented as XXXX-YYYY are coded as being collected in year YYYY.

Source

https://www.shotsforschool.org/k-12/reporting-data/